+++ ToDo/app/src/main/java/com/app4web/asdzendo/todo/launcher/ToDoInjectorUtils.kt
@@ -45,6 +45,6 @@ object ToDoInjectorUtils {
     fun provideToDoViewModelFactory(context: Context): ToDoViewModelFactory =
         ToDoViewModelFactory(getFactRepository(context))
 
-    fun provideFactDetailViewModelFactory(context: Context, factId: Int, paemi: String): FactDetailViewModelFactory =
+    fun provideFactDetailViewModelFactory(context: Context, factId: Long, paemi: String): FactDetailViewModelFactory =
         FactDetailViewModelFactory(getFactRepository(context), factId, paemi)
 }
